feat: input 분리 및 스타일 수정
